requirements.txt
python-telegram-bot==21.6
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
matplotlib==3.8.4
APScheduler==3.10.4
pydub==0.25.1
SpeechRecognition==3.10.4
openai>=1.40.0
